requirements.txt
aiogram
telethon
loguru
python-dotenv
aiofiles
